Moved the alert dialogs to a new file
