+++ src/windows/mod.rs
@@ -1,2 +1,4 @@
+pub mod alert_dialogs;
 pub mod file_handling;
 pub mod generation;
+pub mod regeneration;
