+++ src/windows/regeneration.rs
@@ -1,0 +1,3 @@
+use crate::GtkTestWindow;
+
+impl GtkTestWindow {}
